Annotate test errors by the current sample
In the context of this workflow, we want to compare errors between
different views on the tip attributes for a given timepoint. These views
depend on the sample being analyzed, so we need to annotate the final
errors by the current sample and not the validation sample. We needed
the annotation of the validation sample in the original workflow, since
we needed to compare train/validate/test errors for the same sample.