Fix 16-AA offset in Shih epitope sites
Fixes a bug with the Shih epitope sites that I previously fixed for the
Luksza and Lassig 2014 sites [1] and did not recognize back then. The bug
is a 16 amino acid offset in the HA1 coordinates caused by not properly
accounting for the 16-AA signal peptide when I originally converted the
bitmap implementations of these sites to the distance map JSONs. In the
process of fixing this bug, I also fixed a separate bug where position
HA1:202 was originally included in the map when HA1:201 should have been
included instead. This commit fixes those bugs and adds a citation link
for the Shih et al. sites.

[1] https://github.com/blab/flu-forecasting/commit/4a93171fe98a95af83997f4fbd8832d6b795e17d