Use augur frequencies output for tip attributes
The old workflow used two different approaches to estimate frequencies for
historical reasons that I don't fully remember. This means some parts of the
workflow used augur frequencies (e.g., the tip frequencies display in Auspice)
and other parts used a custom script I wrote at the beginning of PhD that
eventually turned into augur frequencies (e.g., the rest of the workflow). It
turns out that the custom script does not always include the max date as a
pivot which is a similar issue to one I fixed in Augur 20.0.0 [1]. Instead
of maintaining two different approaches in this workflow, I decided to always
use the augur frequencies output for my frequency estimates.

This commit fixes the initial issue with incorrect pivots in the custom frequencies
script by changing the rule for "frequencies_to_table" to use an augur frequencies
JSON instead of the custom script's JSON and by changing the frequencies to table
script to use that augur JSON as input. Subsequent commits will fix other
rules/scripts that depend on the custom frequencies script.

[1] https://github.com/nextstrain/augur/blob/master/CHANGES.md#2000-20-january-2023