Assign clade labels to all nodes in the tree
Changes the logic for clade assignment such that all nodes in the tree get a
clade label based on their HA amino acid sequences. The original implementation
only assigned clade labels to internal nodes and then propagated those labels
to the terminal nodes. The issue with this approach is that a single tip could
be assigned a different clade depending on the other sequences included in the
analysis and this outcome violates a key assumption of several aspects of this
project's clade frequency analysis.