Keep tip AA sequences and not clade membership
We generate clade membership per tip from the full tree for each
dataset, so we no longer have a clade value in the tip attributes table.
However, we want to use the full amino acid sequence to find the most
genetically similar tips between timepoints, so we include that column
in the final output now.