Calculate effects of horizons and delays on optimal distances to the future
Use the optimal distance to the future between populations (the lower
bound on the earth mover's distance between any two populations) to
visualize the effects of reducing forecast horizons and submission
delays. These distances reflect the empirical fitness of each initial
population and are therefore not biased like the distances based on the
estimated fitness.

Closes #15