Update plots and tables for forecast errors
Updates plots and tables for forecast frequency errors to reflect the
use of nested clade frequencies. This commit includes some major
aesthetic changes to account for the increased number of points on most
plots including switch from boxplots to violinplots with quartile lines
and reduced alpha for each point. As part of these changes, this commit
updates the notebook Conda environment to use the latest version of
seaborn (which provides nicer defaults for violinplot quartiles) and
the Jupyter notebook formatter.