Remove unnecessary supplemental figures
Removes supplemental figures that are no longer necessary now that we
calculate clade frequencies as nested values based on each clade and its
descendants. The patterns that we observed with unnested clades
disappear with this switch to nested clades. I suspect that the
systematic bias toward overestimation of large clades that these
supplemental figures were trying to explain actually stemmed from the
use of unnested clade frequencies.