Use optimal forecast frequencies for primary results
Updates popcast to the latest version which exports the optimal projected
frequency per timepoint and strain as part of the calculatio of the optimal
distance between timepoints and then updates the scripts in this workflow to
export that optimal frequency field in the summary TSVs for clade
frequencies. This commit includes the resulting updated clade frequency files
for both natural and simulated populations.

Closes #25