Use absolute forecast errors for main figure
Adds figures for absolute forecast errors for natural and simulated
populations, uses the absolute forecast errors figure for natural
populations as the main figure 4, and moves the other original figures
of overall error distributions to supplemental figures. Updates the main
results text to focus on absolute forecast errors first as a measure of
forecast accuracy followed by a section on the overall errors as a
measure of forecast bias.

Closes #22