Emphasize the primary goal of this study
Clarifies the introduction, discussion, and abstract to emphasize the
primary goal for this study which is to test effects of realistic
changes to public health policy and not to test methodological
improvements to forecasting models. These changes include explcitly
stating that the effects we're testing depend on public health policy
and explaining the focus on sequence-based forecasting models instead of
newer models that include phenotypic data.

Closes #33